requirements.txt
flask
gunicorn
selenium
webdriver-manager
Pillow
requests
